+++ Attendance Management System/Attendance Management System.csproj
@@ -16,9 +16,6 @@
 
   <ItemGroup>
     <Compile Update="Forms\Admin.cs" />
-    <Compile Update="Forms\search.cs">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Compile>
     <Compile Update="Forms\Setting.cs" />
     <Compile Update="Properties\Resources.Designer.cs">
       <DesignTime>True</DesignTime>
