+++ Attendance Management System/Attendance Management System.csproj.user
@@ -13,9 +13,6 @@
     <Compile Update="Forms\AttendencdeReport.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Update="Forms\ChangePassword.cs">
-      <SubType>Form</SubType>
-    </Compile>
     <Compile Update="Forms\Edit.cs">
       <SubType>UserControl</SubType>
     </Compile>
@@ -24,9 +21,6 @@
     </Compile>
     <Compile Update="Forms\Login.cs">
       <SubType>Form</SubType>
-    </Compile>
-    <Compile Update="Forms\search.cs">
-      <SubType>UserControl</SubType>
     </Compile>
     <Compile Update="Forms\Setting.cs">
       <SubType>Form</SubType>
