--- Attendance Management System/Classes/DataClasses/ClassesDataXML.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Attendance_Management_System.Classes.DataClasses
-{
-    public class ClassesDataXML
-    {
-        public string ID {get; set; }   
-        public string Name { get; set; }
-    }
-}
