--- Attendance Management System/Classes/DataClasses/TeacherDataXml.cs
@@ -1,10 +1,0 @@
-﻿namespace Attendance_Management_System.Classes.DataClasses
-{
-    public class TeacherDataXml
-    {
-        public string TeacherId { get; set; }
-        public string Name { get; set; }
-
-        public List<string>? Teacher_Courses { get; set; }
-    }
-}
