+++ Attendance Management System/Forms/Teacher.cs
@@ -3,7 +3,6 @@ using Attendance_Management_System.Forms;
 using System.CodeDom.Compiler;
 using System.Drawing.Text;
 using System.Globalization;
-using System.Linq;
 using System.Windows.Forms;
 
 namespace Attendance_Management_System
@@ -56,8 +55,7 @@ namespace Attendance_Management_System
 
         private void EditPasswordButton_Click(object sender, EventArgs e)
         {
-            ChangePassword formChangePassword = new ChangePassword(session);
-            formChangePassword.Show();
+
         }
 
         private void ViewAttendanceButton_Click(object sender, EventArgs e)
@@ -78,31 +76,7 @@ namespace Attendance_Management_System
 
         private void searchButton_Click(object sender, EventArgs e)
         {
-            SearchPanel.Visible = true;
-            search_Load();
-        }
-        private void search_Load()
-        {
-            dataGridView2.Rows.Clear();
 
-            var ClassesList = DataXMLManagement.GetClassesData();
-            foreach (var classes in ClassesList)
-            {
-                classcombBox.Items.Add(classes.ID);
-            }
-
-
-            var StudentList = DataXMLManagement.GetStudentData();
-            foreach (var student in StudentList)
-            {
-                dataGridView2.Rows.Add(student.studentId, student.name);
-            }
-
-            var ClassList = DataXMLManagement.GetClassesData();
-            foreach (var clas in ClassList)
-            {
-                dataGridView3.Rows.Add(clas.ID, clas.Name);
-            }
         }
 
         public void refresh()
@@ -123,8 +97,6 @@ namespace Attendance_Management_System
             TakeAttendance takeAttendance = new TakeAttendance(session);
             ViewPanel.Controls.Add(takeAttendance);
             ViewPanel.Size = new Size(1502, 900);
-            SearchPanel.Visible = false;
-
         }
 
         private void TakeAttendance_Load(object sender, EventArgs e)
@@ -143,45 +115,6 @@ namespace Attendance_Management_System
             TeacherReport teacherReport = new TeacherReport(session);
             ViewPanel.Controls.Add(teacherReport);
             ViewPanel.Size = new Size(1502, 900);
-            SearchPanel.Visible = false;
-        }
-
-        private void studentSearch_Click(object sender, EventArgs e)
-        {
-            dataGridView2.Rows.Clear();
-
-            var StudentList = DataXMLManagement.GetStudentData();
-
-
-            foreach (var student in StudentList)
-            {
-                if (
-                    (student.studentId.Contains(IdStudenttextBox.Text) || IdStudenttextBox.Text == "") &&
-                    (student.name.Contains(nameStudenttextBox.Text) || nameStudenttextBox.Text == "") &&
-                    (student.Classes.Any(c => c == classcombBox.Text || classcombBox.Text == "") &&
-                    user.Coursesid.Any(Tc => student.Classes.Any(c => c == Tc))
-                    ))
-                    dataGridView2.Rows.Add(student.studentId, student.name);
-
-            }
-        }
-
-        private void ClassSearch_Click(object sender, EventArgs e)
-        {
-            dataGridView3.Rows.Clear();
-
-            var ClassesList = DataXMLManagement.GetClassesData();
-
-
-            foreach (var clas in ClassesList)
-            {
-                if (
-                    (clas.ID.Contains(textBox3.Text) || textBox3.Text == "") &&
-                    (clas.Name.Contains(textBox4.Text) || textBox4.Text == "") &&
-                    user.Coursesid.Any(c => c == clas.ID)
-                    )
-                    dataGridView3.Rows.Add(clas.ID, clas.Name);
-            }
         }
     }
 }
